mangosaur/requirements.txt
Django==2.1.7
django-summernote==0.8.11.4
django-tinymce==2.8.0
django-vote==2.1.7
djangorestframework==3.9.2
djangorestframework-simplejwt==4.1.0
Pillow==5.4.1
psycopg2-binary==2.7.7
PyJWT==1.7.1
pytz==2018.9